requirements.txt
streamlit==1.36.0
snowflake-connector-python==3.11.0
python-dotenv==1.0.1
pandas==2.2.2
httpx==0.27.2
groq==0.9.0
cryptography==42.0.8
PyYAML==6.0.1
anthropic>=0.28.0
PyMuPDF>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0